Fix warning in "Errorf" arguments for Static Analysis Win64.
#rb Rolando.Caloca
#rnx


#ROBOMERGE-SOURCE: CL 13601401 via CL 13601417
#ROBOMERGE-BOT: (v703-13590297)

[CL 13601424 by lukas hermanns in Release-Engine-Staging branch]

+++ Engine/Source/Runtime/Engine/Private/Materials/HLSLMaterialTranslator.cpp
@@ -6867,7 +6867,7 @@ int32 FHLSLMaterialTranslator::CustomExpression( class UMaterialExpressionCustom
 				InputParamDecl += OutputNameStr;
 				break;
 			default:
-				return Errorf(TEXT("Bad type %d for %s output %s"), CustomOutput.OutputType, *Custom->Description, *OutputNameStr);
+				return Errorf(TEXT("Bad type %d for %s output %s"), static_cast<int32>(CustomOutput.OutputType), *Custom->Description, *OutputNameStr);
 				break;
 			}
 		}
